+++ tukaan/_utils.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import collections
 import collections.abc
 import itertools
-from typing import TYPE_CHECKING, Any, Callable, DefaultDict, Iterator, TypeVar
+from typing import TYPE_CHECKING, Any, Callable, DefaultDict, Iterator, Sequence, TypeVar
 
 if TYPE_CHECKING:
     from PIL import Image  # type: ignore
@@ -21,7 +21,7 @@ T_contra = TypeVar("T_contra", contravariant=True)
 
 
 class count:
-    """Simplified itertools.count."""
+    """Simplified itertools.count, that can be int()-ed."""
 
     def __init__(self, start: int = 0) -> None:
         self._count = start
@@ -53,9 +53,32 @@ _widgets: dict[str, TkWidget] = {}
 flatten = itertools.chain.from_iterable
 
 
-def reversed_dict(dictionary: dict) -> dict:
+def reversed_dict(dictionary: dict[object, object]) -> dict[object, object]:
     return {value: key for key, value in dictionary.items()}
 
 
-def seq_pairs(sequence):
+def seq_pairs(sequence: Sequence[object]) -> Iterator[object]:
     return zip(sequence[::2], sequence[1::2])
+
+
+class instanceclassmethod:
+    def __init__(self, method) -> None:
+        self.method = method
+
+    def __get__(self, obj: object | None, objtype: type):
+        def wrapper(*args, **kwargs):
+            if obj is None:
+                return self.method(objtype, *args, **kwargs)
+            return self.method(obj, *args, **kwargs)
+
+        return wrapper
+
+
+class classproperty:
+    def __init__(self, fget):
+        if not isinstance(fget, classmethod):
+            fget = classmethod(fget)
+        self.fget = fget
+
+    def __get__(self, obj: object | None, objtype: type):
+        return self.fget.__get__(obj, objtype)()
