--- tukaan/themes/__init__.py
@@ -1,1 +1,0 @@
-from .themes import AquaTheme, ClamTheme, Theme, Win32Theme, native_theme
